Split files and add test case
