+++ RMVtransport/__init__.py
@@ -1,3 +1,3 @@
 """Define module-level imports."""
 # pylint: disable=C0103
-from RMVtransport.rmvtransport import RMVtransport  # noqa
+from .rmvtransport import RMVtransport  # noqa
