+++ fixtures/request_no_journeys.xml
@@ -18,8 +18,6 @@
       <StartT time="23:17" date="20180830"/>
       <EndT time="23:32" date="20180830"/>
     </SBReq>
-    <JourneyList>
-    </JourneyList>
     <NextUpdate time="23:18" date="20180830"/>
   </SBRes>
 </ResC>
