+++ fixtures/result_midnight.json
@@ -1,0 +1,18 @@
+{
+  "station": "Mainz Hauptbahnhof",
+  "stationId": "3006904",
+  "filter": "11111111111",
+  "journeys": [{
+    "product": "Bus",
+    "number": 54,
+    "trainId": "2746308001",
+    "direction": "Ginsheim-Gustavsburg-Ginsheim Friedr.-Ebert-Platz",
+    "departure_time": "2018-08-31T00:16:00",
+    "minutes": 59,
+    "delay": 0,
+    "stops": ["Mainz Münsterplatz"],
+    "info": "Bischofsheim: Busse 54, 55, 72, 81, 87, 91 - Änderungen vom 23.08. bis 07.09.",
+    "info_long": "Die Busse fahren eine Umleitung. Die Änderungen sind größtenteils in der Verbindungsauskunft enthalten.",
+    "icon": "https://www.rmv.de/auskunft/s/n/img/products/64_pic.png"
+  }]
+}
