+++ tox.ini
@@ -22,3 +22,10 @@ deps = pipenv
 commands=
     make init
     make lint
+
+[testenv:typing]
+whitelist_externals = make
+deps = pipenv
+commands=
+    make init
+    make typing
